All functionality works +/- well. Design is disguisting. May need to fix some bugs in the future
